Start to eliminate the global Order model, replace with: - OrderPlaced event from portfolio.api
We've been a bit lazy in over using a common class that contains a portfolio id, order id, and order details. In some cases we end up with redundant information (e.g. within a persistent entity in which either the portfolio id or the order id are redundant with entityId()). We should disambiguate the OrderPlaced domain event from aggregate events and other uses of this packet of information.