Dockerfile for the UI Reactive platform orchestration setup for Play/Lagom components
Currently not working when tested with 'deploy minikube' due to "[error, com.lightbend.rp.reactivecli.Main$] Failed to obtain Docker config for wiretransferimpl:1.0-SNAPSHOT, unable to access repository or registry; check authentication [401]". Seems to either not be looking at the local docker image repo, or not able to access it.